==> case_study/R/R_diffStation/0_preprocessData_allProcesses.R <==
calibrMod <- 'calibrated'      # calibrated uncalibrated
source('0_preprocessData_discharge.R')
source('0_preprocessData_predictor.R')
source('0_preprocessData_date.R')
source('0_preprocessData_stationInfo.R')
